requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.1